+++ layers/CMSSchema/packages/custompost-category-mutations/src/MutationResolvers/AbstractSetCategoriesOnCustomPostMutationResolver.php
@@ -89,10 +89,23 @@ abstract class AbstractSetCategoriesOnCustomPostMutationResolver extends Abstrac
         ObjectTypeFieldResolutionFeedbackStore $objectTypeFieldResolutionFeedbackStore,
     ): mixed {
         $customPostID = $fieldDataAccessor->getValue(MutationInputProperties::CUSTOMPOST_ID);
-        $customPostCategoryIDs = $fieldDataAccessor->getValue(MutationInputProperties::CATEGORY_IDS);
+        
+        /** @var stdClass|null */
+        $categoriesBy = $fieldDataAccessor->getValue(MutationInputProperties::CATEGORIES_BY);
+        if ($categoriesBy === null || ((array) $categoriesBy) === []) {
+            return $customPostID;
+        }
+        
         $append = $fieldDataAccessor->getValue(MutationInputProperties::APPEND);
-        $customPostCategoryTypeAPI = $this->getCustomPostCategoryTypeMutationAPI();
-        $customPostCategoryTypeAPI->setCategories($customPostID, $customPostCategoryIDs, $append);
+        if (isset($categoriesBy->{MutationInputProperties::IDS})) {
+            /** @var array<string|int> */
+            $customPostCategoryIDs = $categoriesBy->{MutationInputProperties::IDS};
+            $this->getCustomPostCategoryTypeMutationAPI()->setCategoriesByID($customPostID, $customPostCategoryIDs, $append);
+        } elseif (isset($categoriesBy->{MutationInputProperties::SLUGS})) {
+            /** @var string[] */
+            $customPostCategorySlugs = $categoriesBy->{MutationInputProperties::SLUGS};
+            $this->getCustomPostCategoryTypeMutationAPI()->setCategoriesBySlug($customPostID, $customPostCategorySlugs, $append);
+        }
         return $customPostID;
     }
 
@@ -116,12 +129,31 @@ abstract class AbstractSetCategoriesOnCustomPostMutationResolver extends Abstrac
             $objectTypeFieldResolutionFeedbackStore,
         );
 
-        $customPostCategoryIDs = $fieldDataAccessor->getValue(MutationInputProperties::CATEGORY_IDS);
-        $this->validateCategoriesExist(
-            $customPostCategoryIDs,
-            $fieldDataAccessor,
-            $objectTypeFieldResolutionFeedbackStore,
-        );
+        /** @var stdClass */
+        $categoriesBy = $fieldDataAccessor->getValue(MutationInputProperties::CATEGORIES_BY);
+        if (isset($categoriesBy->{MutationInputProperties::IDS})) {
+            $customPostCategoryIDs = $categoriesBy->{MutationInputProperties::IDS};
+            $this->validateCategoriesByIDExist(
+                $customPostCategoryIDs,
+                $fieldDataAccessor,
+                $objectTypeFieldResolutionFeedbackStore,
+            );
+
+            if ($objectTypeFieldResolutionFeedbackStore->getErrorCount() > $errorCount) {
+                return;
+            }
+        } elseif (isset($categoriesBy->{MutationInputProperties::SLUGS})) {
+            $customPostCategorySlugs = $categoriesBy->{MutationInputProperties::SLUGS};
+            $this->validateCategoriesBySlugExist(
+                $customPostCategorySlugs,
+                $fieldDataAccessor,
+                $objectTypeFieldResolutionFeedbackStore,
+            );
+
+            if ($objectTypeFieldResolutionFeedbackStore->getErrorCount() > $errorCount) {
+                return;
+            }
+        }
 
         if ($objectTypeFieldResolutionFeedbackStore->getErrorCount() > $errorCount) {
             return;