+++ layers/CMSSchema/packages/post-tag-mutations/src/SchemaHooks/PostMutationResolverHookSet.php
@@ -6,8 +6,31 @@ namespace PoPCMSSchema\PostTagMutations\SchemaHooks;
 
 use PoPCMSSchema\CustomPostTagMutations\SchemaHooks\AbstractCustomPostMutationResolverHookSet;
 use PoPCMSSchema\PostMutations\SchemaHooks\PostMutationResolverHookSetTrait;
+use PoPCMSSchema\PostTags\TypeResolvers\ObjectType\PostTagObjectTypeResolver;
+use PoPCMSSchema\Tags\TypeResolvers\ObjectType\TagObjectTypeResolverInterface;
 
 class PostMutationResolverHookSet extends AbstractCustomPostMutationResolverHookSet
 {
     use PostMutationResolverHookSetTrait;
+
+    private ?PostTagObjectTypeResolver $postTagObjectTypeResolver = null;
+
+    final public function setPostTagObjectTypeResolver(PostTagObjectTypeResolver $postTagObjectTypeResolver): void
+    {
+        $this->postTagObjectTypeResolver = $postTagObjectTypeResolver;
+    }
+    final protected function getPostTagObjectTypeResolver(): PostTagObjectTypeResolver
+    {
+        if ($this->postTagObjectTypeResolver === null) {
+            /** @var PostTagObjectTypeResolver */
+            $postTagObjectTypeResolver = $this->instanceManager->getInstance(PostTagObjectTypeResolver::class);
+            $this->postTagObjectTypeResolver = $postTagObjectTypeResolver;
+        }
+        return $this->postTagObjectTypeResolver;
+    }
+
+    protected function getTagTypeResolver(): TagObjectTypeResolverInterface
+    {
+        return $this->getPostTagObjectTypeResolver();
+    }
 }